Add files via upload
The update in streams, branches, and commits to projects, models and versions made a need for an update in the code.

The graphs aren't included... I am not motivated to continue as the rest of the graphs are not part of the scope of what I am learning to build. These updates will be useful for anyone looking to follow through with your work.